backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
langchain-openai
python-dotenv
pymongo
motor